rafactoring login
